requirements.txt
redis==3.5.3
elasticsearch==7.12.0
python-dotenv==0.15.0
